CleanUpComponents: remove old views and replace them with new views
