+++ card/src/main/res/layout/card_view.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2019 Adyen N.V.
+  ~ Copyright (c) 2022 Adyen N.V.
   ~
   ~ This file is open source and available under the MIT license. See the LICENSE file for more info.
   ~
-  ~ Created by ran on 13/3/2019.
+  ~ Created by josephj on 16/9/2022.
   -->
 
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
