+++ giftcard/src/main/java/com/adyen/checkout/giftcard/GiftCardView.kt
@@ -1,11 +1,10 @@
 /*
- * Copyright (c) 2021 Adyen N.V.
+ * Copyright (c) 2022 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by josephj on 13/9/2021.
+ * Created by josephj on 30/9/2022.
  */
-
 package com.adyen.checkout.giftcard
 
 import android.content.Context
@@ -14,60 +13,84 @@ import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
 import android.view.View.OnFocusChangeListener
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.Observer
+import android.widget.LinearLayout
+import com.adyen.checkout.components.base.ComponentDelegate
+import com.adyen.checkout.components.extensions.setLocalizedHintFromStyle
+import com.adyen.checkout.components.ui.ComponentViewNew
 import com.adyen.checkout.components.ui.Validation
-import com.adyen.checkout.components.ui.view.AdyenLinearLayout
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.giftcard.databinding.GiftcardViewBinding
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 
-private val TAG = LogUtil.getTag()
-
-class GiftCardView :
-    AdyenLinearLayout<GiftCardOutputData, GiftCardConfiguration, GiftCardComponentState, GiftCardComponent>,
-    Observer<GiftCardOutputData> {
+internal class GiftCardView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+) :
+    LinearLayout(
+        context,
+        attrs,
+        defStyleAttr
+    ),
+    ComponentViewNew {
 
     private val binding: GiftcardViewBinding = GiftcardViewBinding.inflate(LayoutInflater.from(context), this)
 
-    constructor(context: Context) : super(context) {
-        init()
-    }
+    private lateinit var localizedContext: Context
 
-    constructor(context: Context, attributeSet: AttributeSet) : super(context, attributeSet) {
-        init()
-    }
+    private lateinit var giftCardDelegate: GiftCardDelegate
 
-    constructor(context: Context, attributeSet: AttributeSet, defStyleAttr: Int) : super(
-        context,
-        attributeSet,
-        defStyleAttr
-    ) {
-        init()
-    }
-
-    private fun init() {
+    init {
         orientation = VERTICAL
         val padding = resources.getDimension(R.dimen.standard_margin).toInt()
         setPadding(padding, padding, padding, 0)
     }
 
-    override fun initLocalizedStrings(localizedContext: Context) {
-        binding.textInputLayoutGiftcardNumber.setLocalizedHintFromStyle(
-            R.style.AdyenCheckout_GiftCard_GiftCardNumberInput
-        )
-        binding.textInputLayoutGiftcardPin.setLocalizedHintFromStyle(R.style.AdyenCheckout_GiftCard_GiftCardPinInput)
+    override fun initView(delegate: ComponentDelegate, coroutineScope: CoroutineScope, localizedContext: Context) {
+        if (delegate !is GiftCardDelegate) throw IllegalArgumentException("Unsupported delegate type")
+        giftCardDelegate = delegate
+
+        this.localizedContext = localizedContext
+        initLocalizedStrings(localizedContext)
+
+        observeDelegate(delegate, coroutineScope)
+        initInputs()
     }
 
-    override fun initView() {
+    private fun initLocalizedStrings(localizedContext: Context) {
+        binding.textInputLayoutGiftcardNumber.setLocalizedHintFromStyle(
+            R.style.AdyenCheckout_GiftCard_GiftCardNumberInput,
+            localizedContext
+        )
+        binding.textInputLayoutGiftcardPin.setLocalizedHintFromStyle(
+            R.style.AdyenCheckout_GiftCard_GiftCardPinInput,
+            localizedContext
+        )
+    }
+
+    private fun observeDelegate(delegate: GiftCardDelegate, coroutineScope: CoroutineScope) {
+        delegate.outputDataFlow
+            .onEach { outputDataChanged(it) }
+            .launchIn(coroutineScope)
+    }
+
+    private fun outputDataChanged(giftCardOutputData: GiftCardOutputData?) {
+        Logger.v(TAG, "GiftCardOutputData changed")
+        // no ops
+    }
+
+    private fun initInputs() {
         binding.editTextGiftcardNumber.setOnChangeListener {
-            component.inputData.cardNumber = binding.editTextGiftcardNumber.rawValue
+            giftCardDelegate.inputData.cardNumber = binding.editTextGiftcardNumber.rawValue
             notifyInputDataChanged()
             binding.textInputLayoutGiftcardNumber.error = null
         }
 
         binding.editTextGiftcardNumber.onFocusChangeListener = OnFocusChangeListener { _: View?, hasFocus: Boolean ->
-            val cardNumberValidation = component.outputData?.giftcardNumberFieldState?.validation
+            val cardNumberValidation = giftCardDelegate.outputData?.giftcardNumberFieldState?.validation
             if (hasFocus) {
                 binding.textInputLayoutGiftcardNumber.error = null
             } else if (cardNumberValidation != null && cardNumberValidation is Validation.Invalid) {
@@ -76,41 +99,26 @@ class GiftCardView :
         }
 
         binding.editTextGiftcardPin.setOnChangeListener { editable: Editable ->
-            component.inputData.pin = editable.toString()
+            giftCardDelegate.inputData.pin = editable.toString()
             notifyInputDataChanged()
             binding.textInputLayoutGiftcardPin.error = null
         }
 
         binding.editTextGiftcardPin.onFocusChangeListener = OnFocusChangeListener { _, hasFocus ->
-            val pinValidation = component.outputData?.giftcardPinFieldState?.validation
+            val pinValidation = giftCardDelegate.outputData?.giftcardPinFieldState?.validation
             if (hasFocus) {
                 binding.textInputLayoutGiftcardPin.error = null
             } else if (pinValidation != null && pinValidation is Validation.Invalid) {
                 binding.textInputLayoutGiftcardPin.error = localizedContext.getString(pinValidation.reason)
             }
         }
-
-        notifyInputDataChanged()
     }
 
-    override fun observeComponentChanges(lifecycleOwner: LifecycleOwner) {
-        component.observeOutputData(lifecycleOwner, this)
-    }
-
-    override fun onComponentAttached() {
-        // nothing to impl
-    }
-
-    override fun onChanged(giftCardOutputData: GiftCardOutputData?) {
-        Logger.v(TAG, "GiftCardOutputData changed")
-    }
-
-    override val isConfirmationRequired: Boolean
-        get() = true
+    override val isConfirmationRequired: Boolean = true
 
     override fun highlightValidationErrors() {
         Logger.d(TAG, "highlightValidationErrors")
-        val outputData = component.outputData ?: return
+        val outputData = giftCardDelegate.outputData ?: return
         var isErrorFocused = false
         val cardNumberValidation = outputData.giftcardNumberFieldState.validation
         if (cardNumberValidation is Validation.Invalid) {
@@ -128,6 +136,12 @@ class GiftCardView :
     }
 
     private fun notifyInputDataChanged() {
-        component.notifyInputDataChanged()
+        giftCardDelegate.onInputDataChanged(giftCardDelegate.inputData)
+    }
+
+    override fun getView(): View = this
+
+    companion object {
+        private val TAG = LogUtil.getTag()
     }
 }
