+++ issuer-list/src/main/java/com/adyen/checkout/issuerlist/IssuerListRecyclerView.kt
@@ -22,7 +22,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
-class IssuerListRecyclerView @JvmOverloads constructor(
+internal class IssuerListRecyclerView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
