+++ issuer-list/src/main/java/com/adyen/checkout/issuerlist/IssuerListSpinnerView.kt
@@ -24,7 +24,7 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @Suppress("TooManyFunctions")
-class IssuerListSpinnerView @JvmOverloads constructor(
+internal class IssuerListSpinnerView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
