+++ ui-core/src/main/java/com/adyen/checkout/components/ui/PaymentInProgressView.kt
@@ -12,6 +12,7 @@ import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
+import androidx.annotation.RestrictTo
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.adyen.checkout.components.base.ActionDelegate
 import com.adyen.checkout.components.base.ComponentDelegate
@@ -19,6 +20,7 @@ import com.adyen.checkout.components.ui.databinding.ViewPaymentInProgressBinding
 import com.adyen.checkout.core.exception.CancellationException
 import kotlinx.coroutines.CoroutineScope
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class PaymentInProgressView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
